# packages/contracts/openapi/runtime-gateway.openapi.yaml
openapi: 3.1.0
info:
  title: AgentTwin runtime gateway API
  version: 1.0.0
  summary: Runtime containment of an agent's tool actions — policies, approvals and every decision.
  description: |
    The runtime gateway is an opt-in path through which an agent's tool
    actions pass at runtime (spec §30-31, ADR-0033). Each call is decided by
    the tool's active policies: allowed, allowed with limits, sent to a
    person for approval, or denied — before it reaches the tool.

    * **Tool endpoints** name the tools the gateway may call, their risk tier
      and timeout; a URL outside the egress allowlist is refused.
    * **Policies** are named, versioned documents of CEL rules that guard one
      tool. A version is tested (its own tests and every numeric threshold)
      and activated only when every test passes.
    * **Approvals** bind a person's decision to one exact action: the approved
      action runs once; a changed action needs its own approval.
    * **Decisions** record every call the gateway decided (append-only).

    **Where to call it.** The public API is served by the control plane (the
    edge), which authenticates the caller, applies rate limits and, on
    `/api/v1` mutations, `Idempotency-Key`; it forwards the request to the
    runtime gateway. The invocation path (`/gateway/v1`) is reached the same
    way with an API key holding the `runtime:invoke` scope; there the
    gateway itself implements `Idempotency-Key`, per action, and forwards the
...

...

...
servers:
...
security:
  - bearerAuth: []
  - apiKey: []
paths:
...
  /api/v1/approvals:
    get:
      operationId: listApprovals
      tags: [Approvals]
      summary: List approval requests
      description: |
        Newest first. `status` is the status as of now: `EXPIRED` includes
        open requests past their expiry, which `PENDING` and `APPROVED`
        exclude.
      parameters:
        - $ref: '#/components/parameters/Organization'
        - $ref: '#/components/parameters/Project'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ApprovalStatus'
        - name: tool
          in: query
          schema:
            $ref: '#/components/schemas/ToolName'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: A page of requests.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
        '502':
          $ref: '#/components/responses/UpstreamUnavailable'
        '503':
          $ref: '#/components/responses/Unavailable'
  /api/v1/approvals/{approval_id}:
    parameters:
      - $ref: '#/components/parameters/ApprovalId'
    get:
      operationId: getApproval
      tags: [Approvals]
      summary: Read an approval request
      description: |
        What the agent is trying to do (the tool, its risk tier, the redacted
        arguments), why it needs approval (the policy, rule and reason), the
        decision that asked, the expiry, the trace, and every call that
        presented the approval's token with how its arguments differ from the
        approved ones (spec §41.9).
      parameters:
        - $ref: '#/components/parameters/Organization'
        - $ref: '#/components/parameters/Project'
      responses:
        '200':
          description: The request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
        '502':
          $ref: '#/components/responses/UpstreamUnavailable'
        '503':
          $ref: '#/components/responses/Unavailable'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: A session token (dev login), an OIDC access token or an API key.
    apiKey:
      type: apiKey
      in: header
      name: X-AgentTwin-Api-Key
      description: A project API key.
  parameters:
...
    Organization:
      name: X-AgentTwin-Org
      in: header
      required: false
      description: The organization to act in, for credentials that belong to several.
      schema:
        $ref: '#/components/schemas/Uuid'
    Project:
      name: project_id
      in: query
      required: true
      description: The project. A project outside the caller's access answers `404`.
      schema:
        $ref: '#/components/schemas/Uuid'
...
    ApprovalId:
      name: approval_id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/Uuid'
    Cursor:
      name: cursor
      in: query
      description: The `next_cursor` of the previous page.
      schema:
        type: string
        maxLength: 512
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
...
  responses:
    BadRequest:
      description: |
        The request is invalid: `INVALID_PARAMETER` (`details.field`),
        `INVALID_CURSOR`, `PROJECT_REQUIRED` or `INVALID_IDEMPOTENCY_KEY`.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Unauthenticated:
      description: No valid credential (`UNAUTHENTICATED`).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: The credential lacks the permission (`FORBIDDEN`; `details.required_permission`).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: The project or resource does not exist or is outside the caller's projects (`NOT_FOUND`).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    RateLimited:
      description: Too many requests (`RATE_LIMITED`).
      headers:
        Retry-After:
          description: Seconds to wait.
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Internal:
      description: An unexpected error (`INTERNAL`); quote the `request_id`.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    UpstreamUnavailable:
      description: The runtime gateway is not reachable from the control plane (`UPSTREAM_UNAVAILABLE`).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unavailable:
      description: |
        The runtime gateway is not configured on this control plane
        (`SERVICE_NOT_CONFIGURED`) or its database is unavailable
        (`UNAVAILABLE`).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Uuid:
      type: string
      format: uuid
    Timestamp:
      type: string
      format: date-time
    ToolName:
      type: string
      pattern: ^[A-Za-z0-9][A-Za-z0-9_.:-]{0,127}$
    TraceId:
      type: string
      pattern: ^[0-9a-f]{32}$
    Hash:
      type: string
      pattern: ^[0-9a-f]{64}$
    Risk:
      type: string
      enum: [READ, WRITE_REVERSIBLE, WRITE_IRREVERSIBLE, EXECUTE, ADMIN]
    Effect:
      type: string
      enum: [allow, allow_with_limits, require_approval, deny]
    Outcome:
      type: string
      description: |
        `forwarded` (sent, not yet answered), `executed` (the tool answered),
        `failed` (it failed, timed out or answered 5xx), `replayed` (a stored
        answer), `denied`, `approval_required`, `approval_refused` (a token
        that did not let the call run).
      enum: [forwarded, executed, failed, replayed, denied, approval_required, approval_refused]
    ApprovalStatus:
      type: string
      enum: [PENDING, APPROVED, DENIED, EXPIRED, USED]
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              type: string
            message:
              type: string
            request_id:
              type: string
            details:
              type: object
...
    Limits:
      type: object
      properties:
        timeoutMs:
          type: integer
        maxResponseBytes:
          type: integer
    Match:
      type: object
      required: [rule, effect]
      properties:
        rule:
          type: string
        effect:
          $ref: '#/components/schemas/Effect'
        message:
          type: string
    PolicyDecision:
      type: object
      required: [policy, effect, message, matched, fail_mode_applied]
      properties:
        policy:
          type: string
        effect:
          $ref: '#/components/schemas/Effect'
        rule:
          type: string
          description: The rule that decided; absent when the default applied.
        message:
          type: string
        matched:
          type: array
          description: Every rule whose condition held, in order.
          items:
            $ref: '#/components/schemas/Match'
        errors:
          type: array
          description: Rules that could not be evaluated (the fail mode applied).
          items:
            type: object
            required: [rule, error]
            properties:
              rule:
                type: string
              error:
                type: string
        fail_mode_applied:
          type: boolean
        limits:
          $ref: '#/components/schemas/Limits'
...
    Approval:
      type: object
      required: [id, project_id, tool, risk, agent, agent_version, environment, trace_id, action_hash, arguments,
        summary, policy, policy_version_id, rule, reason, decision_id, status, requested_by, expires_at, decided_by,
        decided_at, decision_reason, used_at, used_decision_id, created_at, updated_at]
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        project_id:
          $ref: '#/components/schemas/Uuid'
        tool:
          $ref: '#/components/schemas/ToolName'
        risk:
          $ref: '#/components/schemas/Risk'
        agent:
          type: string
        agent_version:
          type: string
        environment:
          type: string
        trace_id:
          oneOf:
            - $ref: '#/components/schemas/TraceId'
            - type: 'null'
        action_hash:
          $ref: '#/components/schemas/Hash'
        arguments:
          type: object
          description: The action's arguments, redacted.
        summary:
          type: string
          description: A one-line description of the action.
        policy:
          type: string
        policy_version_id:
          oneOf:
            - $ref: '#/components/schemas/Uuid'
            - type: 'null'
        rule:
          type: string
        reason:
          type: string
          description: Why the action needs approval (the rule's message).
        decision_id:
          $ref: '#/components/schemas/Uuid'
        status:
          $ref: '#/components/schemas/ApprovalStatus'
        requested_by:
          type: string
        expires_at:
          $ref: '#/components/schemas/Timestamp'
        decided_by:
          type: [string, 'null']
        decided_at:
          oneOf:
            - $ref: '#/components/schemas/Timestamp'
            - type: 'null'
        decision_reason:
          type: [string, 'null']
        used_at:
          oneOf:
            - $ref: '#/components/schemas/Timestamp'
            - type: 'null'
        used_decision_id:
          oneOf:
            - $ref: '#/components/schemas/Uuid'
            - type: 'null'
        created_at:
          $ref: '#/components/schemas/Timestamp'
        updated_at:
          $ref: '#/components/schemas/Timestamp'
    ApprovalPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Approval'
        next_cursor:
          type: string
    Change:
      type: object
      required: [path]
      properties:
        path:
          type: string
          description: The argument (`amount`, `items[1].sku`).
        before:
          description: The approved value (absent when the argument was not there).
        after:
          description: The attempted value (absent when the argument is gone).
        added:
          type: boolean
        removed:
          type: boolean
    Attempt:
      type: object
      required: [id, result, action_hash, arguments, subject, trace_id, decision_id, created_at, changes]
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        result:
          type: string
          enum: [executed, mismatch, used, expired, not_approved]
        action_hash:
          $ref: '#/components/schemas/Hash'
        arguments:
          type: object
        subject:
          type: string
        trace_id:
          type: [string, 'null']
        decision_id:
          oneOf:
            - $ref: '#/components/schemas/Uuid'
            - type: 'null'
        created_at:
          $ref: '#/components/schemas/Timestamp'
        changes:
          type: array
          description: How the attempted arguments differ from the approved ones.
          items:
            $ref: '#/components/schemas/Change'
    ApprovalDetail:
      allOf:
        - $ref: '#/components/schemas/Approval'
        - type: object
          required: [decision, used_decision, attempts]
          properties:
            decision:
              description: The decision that asked for the approval.
              oneOf:
                - $ref: '#/components/schemas/Decision'
                - type: 'null'
            used_decision:
              description: The decision that executed the approved action.
              oneOf:
                - $ref: '#/components/schemas/Decision'
                - type: 'null'
            attempts:
              type: array
              items:
                $ref: '#/components/schemas/Attempt'
...
    RecordedDecision:
      allOf:
        - $ref: '#/components/schemas/PolicyDecision'
        - type: object
          required: [policy_version_id, version]
          properties:
            policy_version_id:
              $ref: '#/components/schemas/Uuid'
            version:
              type: integer
    Decision:
      type: object
      required: [id, tool, risk, agent, agent_version, environment, subject, trace_id, action_hash, arguments, summary,
        effect, outcome, policy, policy_version_id, rule, message, decisions, fail_mode_applied, approval_id,
        idempotency_key, limits, upstream_status, error_code, latency_ms, created_at, completed_at]
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        tool:
          $ref: '#/components/schemas/ToolName'
        risk:
          $ref: '#/components/schemas/Risk'
        agent:
          type: string
        agent_version:
          type: string
        environment:
          type: string
        subject:
          type: string
          description: Who called (`apikey:<id>`).
        trace_id:
          oneOf:
            - $ref: '#/components/schemas/TraceId'
            - type: 'null'
        action_hash:
          $ref: '#/components/schemas/Hash'
        arguments:
          type: object
          description: The arguments, redacted.
        summary:
          type: string
        effect:
          description: The combined effect; `null` for a replay (decided by the call it replays).
          oneOf:
            - $ref: '#/components/schemas/Effect'
            - type: 'null'
        outcome:
          $ref: '#/components/schemas/Outcome'
        policy:
          type: [string, 'null']
        policy_version_id:
          oneOf:
            - $ref: '#/components/schemas/Uuid'
            - type: 'null'
        rule:
          type: [string, 'null']
        message:
          type: string
        decisions:
          type: array
          description: What each active policy decided.
          items:
            $ref: '#/components/schemas/RecordedDecision'
        fail_mode_applied:
          type: boolean
        approval_id:
          oneOf:
            - $ref: '#/components/schemas/Uuid'
            - type: 'null'
        idempotency_key:
          type: [string, 'null']
        limits:
          oneOf:
            - $ref: '#/components/schemas/Limits'
            - type: 'null'
        upstream_status:
          type: [integer, 'null']
        error_code:
          type: [string, 'null']
        latency_ms:
          type: [integer, 'null']
        created_at:
          $ref: '#/components/schemas/Timestamp'
        completed_at:
          oneOf:
            - $ref: '#/components/schemas/Timestamp'
            - type: 'null'
...